refactor: dashboard as home page.
