+++ apps/codebility/middleware.ts
@@ -108,7 +108,7 @@ function getPatterns() {
       if (!user) return;
       
       // If user is logged in, redirect to their respective page.
-      let nextURL = new URL("/dashboard", req.nextUrl.origin).href; // default for now, mentor.
+      let nextURL = new URL("/home", req.nextUrl.origin).href; // default for now, mentor.
       switch ( userData?.role_type ) {
         case "INTERN":
           nextURL = new URL("/waiting", req.nextUrl.origin).href; // waiting page for now.
